test(react-pi): cover async runtime callbacks
